add lift request tests
